requirements.txt
annotated-types==0.6.0
anyio==4.0.0
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
deprecation==2.1.0
exceptiongroup==1.1.3
Flask==3.0.0
Flask-Testing==0.8.1
gotrue==1.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
numpy==1.26.1
packaging==23.2
pandas==2.1.2
postgrest==0.11.0
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
realtime==1.0.0
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
sniffio==1.3.0
storage3==0.6.1
StrEnum==0.4.15
supabase==1.2.0
supafunc==0.2.3
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
websockets==10.4
Werkzeug==3.0.1
openai==0.12.3
